src/main/java/com/kolosque/loja_de_games/repository/UsuarioRepository.java
package com.kolosque.loja_de_games.repository;

import com.kolosque.loja_de_games.model.Usuario;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.repository.query.Param;
import org.springframework.stereotype.Repository;

import java.util.List;
import java.util.Optional;

@Repository
public interface UsuarioRepository extends JpaRepository<Usuario, Long> {

    public Optional<Usuario> findByEmail(String email);

    public List<Usuario> findAllByNomeContainingIgnoreCase(@Param("nome") String nome);
}
